Move Protractor tests to frontend/tests/integration
See 6ec5ca58.
