+++ frontend/tests/integration/mocks/groups.js
