+++ frontend/tests/integration/mocks/projects.js
