+++ frontend/tests/integration/pages/work-package-details-pane.js
