+++ frontend/tests/integration/work-package-details-spec.js
